Fix error for destructuring default values
Fix https://github.com/arrizalamin/js-function-reflector/issues/17

This commit fixes errors for syntax such as `function([param=5] = [])`
and prevents information on destructuring parameters for syntax such as
`function([param]=[2])` from being lost.

In short, a parameter cannot be destructuring and have a default value
at the same time, the parsed results for parameters looks like this for
the function `function([param=1]) {}`

```javascript
{
params: [
  {
    type: "DESTRUCTURING"
    value: {
      keys: [
        {
          name: "param",
          type: "KEY_WITH_DEFAULT",
          value: 1
        }
      ]
      type: "array"
  }
}
```

and like this for the function `function([param]=[5]) {}`

```javascript
{
  // […]
  params: [
  {
    name: undefined,
    type: "DEFAULT"
    value: [5]
  }
}
```

while the function `function([param=5]=[]) {}` would simply produce an
error. This fix changes it so that the `name` property of a
destructuring parameter with a default value contains the information
which is seen in the `value` property above, while the `value` property
is used for the default value. This way, no information is lost.